Bugfix: Empty database file is never initialized
chrome/content/overlay.js

- (CertWatch.dbinit)
Initialize the database if it's empty.

Prior to this change tables were not created in an empty database file
and as a result the extension would not function. An empty database
file could be created erroneously if a version of the extension with
E4X javascript is used in a version of Firefox that does not support it.